backend/requirements.txt
backoff==2.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coinbase-advanced-py==1.8.2
cryptography==44.0.0
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
urllib3==2.3.0
websockets==13.1
Werkzeug==3.1.3
zipp==3.21.0
